Added some style and extensibility to the templates.
I gave the templates a little style to match the Django colors scheme.  I also
reorganized the templates and changed the JavascriptTestRunner view so that
it points to an overview.html template.  That, in turn, points to an
overview.html template under a django subdirectory that contains templates
that add django's styles to the javascript test runner.  That way, users can
easily override the templates to remove django's color scheme.